chore(app): move files to have better aliases
