+++ packages/app/jest.config.js
@@ -26,6 +26,8 @@ const customJestConfig = {
   testEnvironment: "jest-environment-jsdom",
   moduleNameMapper: {
     "^@/(.*)$": "<rootDir>/src/$1",
+    "^@common/(.*)$": "<rootDir>/src/common/$1",
+    "^@components/(.*)$": "<rootDir>/src/components/$1"
   },
 }
