+++ packages/app/src/common/utils/string.ts
