+++ packages/app/src/hooks/useConfig.ts
@@ -1,10 +1,8 @@
 import useSWRImmutable from "swr/immutable"
 import React from "react"
 
-import type { SelectItemsType } from "../types/utils/select"
-
-import type { FetcherReturnImmutable } from "../utils/fetcher"
-import { fetcher } from "../utils/fetcher"
+import type { FetcherReturnImmutable } from "./utils"
+import { fetcher } from "./utils"
 
 export type ConfigTypeApi = {
   YEARS: number[]
@@ -25,6 +23,9 @@ export type ConfigTypeFormatted = ConfigTypeApi & {
   LAST_PUBLIC_YEAR: string
   PUBLIC_YEARS_TRIES: number[]
 }
+
+type SelectItemsType = [string, string][]
+
 
 /**
  * Return all departments found in config API endpoint, possibly filtered by region.
