+++ packages/app/src/hooks/useStats.ts
@@ -1,7 +1,7 @@
 import useSWR from "swr"
 
-import { fetcher, FetcherReturn } from "../utils/fetcher"
-import { buildUrlParamsString } from "../utils/url"
+import { fetcher, FetcherReturn } from "./utils"
+import { buildUrlParamsString } from "@common/utils/url"
 
 export type StatsType = {
   avg: number
