+++ packages/app/src/pages/index.tsx
@@ -16,9 +16,9 @@ import { useRouter } from "next/router"
 import Head from "next/head"
 import { format } from "date-fns"
 
-import ButtonAction from "../components/ds/ButtonAction"
-import { SinglePageLayout } from "../components/ds/SinglePageLayout"
-import { AverageIndicator } from "../components/AverageIndicator"
+import ButtonAction from "@components/ds/ButtonAction"
+import { SinglePageLayout } from "@components/ds/SinglePageLayout"
+import { AverageIndicator } from "@components/AverageIndicator"
 
 async function getDateCsv(): Promise<string> {
   try {
