+++ packages/app/src/theme/styles/globals.css
