--- packages/app/src/types/utils/select.ts
@@ -1,1 +1,0 @@
-export type SelectItemsType = [string, string][]
